Try to fix CI git error
When attempting to run the pre-commit hooks in the CI, we get this
error:

An error has occurred: FatalError: git failed. Is it installed, and are you in a Git repository directory?

https://github.com/actions/checkout/issues/766 suggests the issue could
be because 'actions/checkout' clones the repository as a different user
than is used in our custom docker container. Git then doesn't trust the
repository and errors when called upon.