Skip no-commit-to-branch pre-commit check in CI
This causes issues when the CI is run on a push to main, as the
no-commit-to-branch check will always fail.

This hook is really to prevent developers committing directly to main
locally. We can safely skip it in our CI, as branch protections will
prevent new commits being pushed to main.