Restore trailing whitespace to OF mesh file
This file is regenerated by OpenFOAM when the test is run, but
OpenFOAM's version contains trailing whitespace. pre-commit won't let us
commit the file with trailing whitespace, so it's stripped. This means
the file is marked as modified after the test is run, which is annoying.

Strip the whitespace for now to make pre-commit happy and we'll deal
with core of the issue later.