Replace buoyantFoam solver with generic solver
OpenFOAM's buoyantFoam solver was removed between versions 10 and 12. It
was replaced with a generic 'fluidSolver'. This commit removes the old
buoyantFoam code and introduces a new class: FoamSolver.

FoamSolver wraps an OpenFOAM 'solver' object and handles running the
timesteps and transferring mesh values (however I have not implemented
this yet!).