Remove FoamInterface class
This class being a singleton was causing problems when more than one
hippo multiapp was being run - it would only run one of them, twice!

This change removes the FoamInterface class and we indirectly replace it
with a FoamRuntime class that keeps everything we need alive (the
Foam::Time, the Foam::argList, etc.). Interactions with the OpenFOAM API
are performed through the rest of the FoamX classes (e.g., FoamSolver,
FoamMesh), so we don't need the interface class anymore.